Fix(Next-Gen Dataset): fix issue causing empty channels when predicting on channel subset (#661)
## Description

<!-- This section provides the necessary background and information for
reviewers to
understand the code and have the correct mindset when examining changes.
-->

> [!NOTE]  
> **tldr**: Fix an issue inserting empty channels in stitching
reconstruction.

This PR does the following:
- Update `data_shape`, which is used to reconstruct the predictions, to
reflect the change in channels and remove singleton dimensions. This
happens in the `CAREamicsDataset` when instantiating `ImageRegionData`.
- Along the way, fixed a wrong parameter in `create_ng_dataset_config`.
- Added docs to `ImageRegionData`.
- Updated tests to also work with channels and, for some, work directly
with the `CAREamicsDataset` to avoid having to update constantly how the
test inputs are generated.


### Background - why do we need this PR?

<!-- What problem are you solving? Describe in a few sentences the state
before
this PR. Use code examples if useful. -->

The `channels` parameter allow selecting a subset of channels for
training/prediction. However, upon reconstructing the data after
prediction, either through stitching or writing Zarr tiles, the original
data size is used and is returned with the missing channels being empty
arrays.

### Overview - what changed?

<!-- What aspects and mechanisms of the code base changed? Describe only
the general
idea and overarching features. -->

`ImageRegionData` carries all the metadata associated with a patch or
tile, including the original `data_shape` used during stitching. This PR
modifies the `data_shape` to reflect the changes in channels.

### Implementation - how did you implement the changes?

<!-- How did you solve the issue technically? Explain why you chose this
approach and
provide code examples if applicable (e.g. change in the API for users).
-->

In `CAREamicsDataset`, when instantiating `ImageRegionData`,
`data_shape` is adjusted.

## Changes Made

<!-- This section highlights the important features and files that
reviewers should
pay attention to when reviewing. Only list important features or files,
this is useful for
reviewers to correctly assess how deeply the modifications impact the
code base.

For instance:

### New features or files
- `NewClass` added to `new_file.py`
- `new_function` added to `existing_file.py`

...
-->

### Modified features or files

<!-- List important modified features or files. -->
- `CAREamicsDataset`: 
- adjust `data_shape` for `ImageRegionData` to reflect actual channels.
    - Added docs to `ImageRegionData`.
- `create_ng_dataset_config`: removed `patch_overlap` parameter which is
not used since it only creates a training configuration (`random`
strategy).


## Related Issues

<!-- Link to any related issues or discussions. Use keywords like
"Fixes", "Resolves",
or "Closes" to link to issues automatically. -->

- Resolves https://github.com/CAREamics/careamics/issues/644

## Notes

This PR uncovered https://github.com/CAREamics/careamics/issues, for
which (failing) tests are implemented (but commented out).

---

**Please ensure your PR meets the following requirements:**

- [x] Code builds and passes tests locally, including doctests
- [x] New tests have been added (for bug fixes/features)
- [x] Pre-commit passes
- [ ] PR to the documentation exists (for bug fixes / features)